refactor: rename `ThingFormationAssociation` file`ThingGeologicFormationAssociation`
