+++ db/thing_geologic_formation_association.py
